feat: expose v2 as default
