--- lua/yasl-v2/default.lua
@@ -1,6 +1,0 @@
-return {
-	global = true,
-	components = {
-		"%t",
-	}
-}
